transaction_service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]
python-multipart
httpx